+++ assets/javascripts/kitten/components/steppers/vertical-stepper/components/title.js
@@ -15,7 +15,6 @@ const StyledTitle = styled.p`
   &.k-Stepepers-VerticalStepper__title--orion {
     font-size: ${stepToRem(-1)};
   }
-
 `
 
 export const Title = ({ variant, className, children }) => {