+++ assets/javascripts/kitten/components/steppers/vertical-stepper/test.js
@@ -69,24 +69,34 @@ describe('<VerticalStepper />', () => {
         .create(
           <VerticalStepper variant="orion">
             <VerticalStepper.Step variant="orion" success href="#">
-              <VerticalStepper.Title variant="orion">Title</VerticalStepper.Title>
+              <VerticalStepper.Title variant="orion">
+                Title
+              </VerticalStepper.Title>
               <VerticalStepper.Link variant="orion">Link</VerticalStepper.Link>
             </VerticalStepper.Step>
 
             <VerticalStepper.Step variant="orion" valid href="#">
-              <VerticalStepper.Title variant="orion">Title</VerticalStepper.Title>
+              <VerticalStepper.Title variant="orion">
+                Title
+              </VerticalStepper.Title>
             </VerticalStepper.Step>
 
             <VerticalStepper.Step variant="orion" error href="#">
-              <VerticalStepper.Title variant="orion">Title</VerticalStepper.Title>
+              <VerticalStepper.Title variant="orion">
+                Title
+              </VerticalStepper.Title>
             </VerticalStepper.Step>
 
             <VerticalStepper.Step variant="orion" waiting href="#">
-              <VerticalStepper.Title variant="orion">Title</VerticalStepper.Title>
+              <VerticalStepper.Title variant="orion">
+                Title
+              </VerticalStepper.Title>
             </VerticalStepper.Step>
 
             <VerticalStepper.Step variant="orion" href="#">
-              <VerticalStepper.Title variant="orion">Title</VerticalStepper.Title>
+              <VerticalStepper.Title variant="orion">
+                Title
+              </VerticalStepper.Title>
             </VerticalStepper.Step>
           </VerticalStepper>,
         )